src/main/java/ru/job4j/kiss/Fool.java
package ru.job4j.kiss;

import java.util.Scanner;

public class Fool {
    public static String getAnswer(int num) {
        String result = String.valueOf(num);
        if (num % 3 == 0 && num % 5 == 0) {
            return "FizzBuzz";
        } else if (num % 3 == 0) {
            return "Fizz";
        } else if (num % 5 == 0) {
            return "Buzz";
        } else {
            return result;
        }
    }

    public static void main(String[] args) {
        System.out.println("Игра FizzBuzz.");
        var startAt = 1;
        var io = new Scanner(System.in);
        while (startAt < 100) {
            System.out.println(Fool.getAnswer(startAt));
            startAt++;
            if (!io.nextLine().equals(Fool.getAnswer(startAt))) {
                System.out.println("Ошибка. Начинай снова.");
                startAt = 0;
            }
            startAt++;
        }
    }
}
